+++ bigquery/CHANGES.md
@@ -8,6 +8,8 @@
 
 1.  Removed unused `mapred.bq.output.buffer.size` configuration property.
 
+1.  Fix unauthenticated access support (`mapred.bq.auth.null.enable=true`).
+
 ### 1.1.1 - 2020-03-11
 
 1.  Fix shaded jar - add back missing relocated dependencies.